+++ src/main/java/at/hannibal2/skyhanni/features/inventory/wardrobe/ArmorWardrobeApi.kt
@@ -24,20 +24,23 @@ object ArmorWardrobeApi : AbstractWardrobeApi() {
 
     override val storage get() = ProfileStorageData.profileSpecific?.wardrobe
 
-    var inCustomWardrobe = false
-
     @HandleEvent
-    fun onInventoryOpen(event: InventoryOpenEvent) {
-        val matched = handleInventoryOpen(event.inventoryName)
-        if (CustomWardrobe.config.enabled) inCustomWardrobe = matched
+    private fun onInventoryOpen(event: InventoryOpenEvent) {
+        handleInventoryOpen(event.inventoryName)
     }
 
     @HandleEvent(priority = HandleEvent.HIGH, onlyOnSkyblock = true)
-    fun onInventoryUpdated(event: InventoryUpdatedEvent) = handleInventoryUpdated(event)
+    private fun onInventoryUpdated(event: InventoryUpdatedEvent) = handleInventoryUpdated(event)
 
     @HandleEvent
-    fun onInventoryClose(event: InventoryCloseEvent) = handleInventoryClose()
+    private fun onInventoryClose(event: InventoryCloseEvent) = handleInventoryClose()
 
     @HandleEvent
-    fun onDebugDataCollect(event: DebugDataCollectEvent) = handleDebugDataCollect(event)
+    private fun onDebugDataCollect(event: DebugDataCollectEvent) = handleDebugDataCollect(event)
+
+    // Sets inThisWardrobe to true if it matches
+    // Called by MenuScreensHook to determine if the inventory should be replaced with the custom wardrobe screen
+    internal fun matchesInventoryName(inventoryName: String): Boolean {
+        return handleInventoryOpen(inventoryName)
+    }
 }
