+++ src/main/java/at/hannibal2/skyhanni/features/inventory/wardrobe/EstimatedWardrobePrice.kt
@@ -8,7 +8,6 @@ import at.hannibal2.skyhanni.events.minecraft.add
 import at.hannibal2.skyhanni.events.minecraft.addAll
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.test.command.ErrorManager
-import at.hannibal2.skyhanni.utils.SkyBlockUtils
 
 @SkyHanniModule
 object EstimatedWardrobePrice {
@@ -16,7 +15,7 @@ object EstimatedWardrobePrice {
     private val config get() = SkyHanniMod.feature.inventory.estimatedItemValues
 
     @HandleEvent(onlyOnSkyblock = true)
-    fun onToolTip(event: ToolTipTextEvent) {
+    private fun onToolTip(event: ToolTipTextEvent) {
         event.slot ?: return
         val api = activeWardrobeApi() ?: return
         val slot = api.slots.firstOrNull {
@@ -31,7 +30,7 @@ object EstimatedWardrobePrice {
         try {
             tooltip.add("")
             tooltip.addAll(lore)
-        } catch (e: IndexOutOfBoundsException) {
+        } catch (_: IndexOutOfBoundsException) {
             ErrorManager.logErrorStateWithData(
                 "Can not show Estimated Wardrobe Price",
                 "IndexOutOfBoundsException while trying to add the estimated wardrobe price line to the tooltip",
@@ -41,8 +40,7 @@ object EstimatedWardrobePrice {
     }
 
     private fun activeWardrobeApi(): AbstractWardrobeApi? {
-        if (!SkyBlockUtils.inSkyBlock) return null
-        if (ArmorWardrobeApi.inCustomWardrobe && !CustomWardrobe.editMode) return null
+        if (CustomWardrobe.inCustomWardrobe) return null
         return when {
             config.armor && ArmorWardrobeApi.inWardrobe() -> ArmorWardrobeApi
             config.equipment && EquipmentWardrobeApi.inWardrobe() -> EquipmentWardrobeApi
@@ -51,7 +49,7 @@ object EstimatedWardrobePrice {
     }
 
     @HandleEvent
-    fun onConfigFix(event: ConfigUpdaterMigrator.ConfigFixEvent) {
+    private fun onConfigFix(event: ConfigUpdaterMigrator.ConfigFixEvent) {
         event.move(3, "misc.estimatedIemValueArmor", "misc.estimatedItemValues.armor")
     }
 }
